feat: update to the new job schema
Problem:
 The job template schema is being updated for service release. With it,
the implementation library is being updated in a breaking way and so the agent
code needs to be updated to that revision.

 Furthermore, this needs to be a phased update. The agent needs to
support both old and new revisions at the same time. Fortunately, there
is nothing changing from the 2022-09-01 revision that is relevant to
the agent except for the version string.

Solution:
 Updated the dependency of the template implementation library to the
relevant version. Also allow the agent to accept both the new and old
revision version string from the service, but treat both values as being
the new revision's version string.

 Some TODOs have been left in the code as markers for the changes that
need to happen in the second phase of the update.

Signed-off-by: Daniel Neilson <neilsd@amazon.com>