fix: crash if queue has no user defined
The fix in https://github.com/casillas2/deadline-cloud-worker-agent/commit/fab7d538d598285d8385779e1b82114565ea0e8f
caused a regression whereby the agent would crash if it dequeues a job
from a Queue that has an empty posix user/group defined. The crash was
due to the QueueBoto3Session being passed a PosixSessionUser with empty
user/group -- it expects to be passed None in this case.

This fixes the issue.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>