feat!: adds compatibility with upcoming BatchGetJobEntity API change (#139)
Problem:
 The contents of the `template` field in the `stepDetails` response to
BatchGetJobEntity API is being changed. It currently contains an Open
Job Description StepScript like so:
```
{
   "embeddedFiles": [ ... ],
   "actions": { "onRun": ... }
}
```

This is being changed to contain an Open Job Description StepTemplate,
like so:
```
{
  "name": <step name>, (max 64 characters)
  "script": {
     "embeddedFiles": [ ... ],
     "actions": { "onRun": ... }
  }
}
```

The API response will *NOT* contain the entirety of the StepTemplate
submitted with the Job, but rather just the subset that is relevant for
the Worker to operate.

Solution:
 This is step 1 in a three phase deployment. It is updating the Worker
Agent to understand both API response shapes, and act accordingly. The
next step, after the service has been updated, will be to delete the
code path related to the old API response shape.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>