fix: no longer sigterm agent when running jobs as same user (#161)
Problem:
When running jobs as the same user as the agent (i.e. the Queue's
jobsRunAs user is the same os user) the agent is SIGTERMing itself when
there is no work left for it to work on. It shouldn't be doing that.

Solution:
We were doing a /usr/bin/whoami to find out who the agent is running as,
but that output had a trailing newline that was interfering with the
check for whether the job user and agent user are the same. This fixes
that.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>