fix: permissions on /var/log/amazon directory
Problem:
It's been reported that when the worker-agent installer, running as
root, creates `/var/log/amazon/deadline` that the agent was unable
to write its own logs to that directory when running as a non-root
user.

Solution:
Add go+rx to `/var/log/amazon` -- this is the solution reported to work
by the reporting customer.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>