feat: report action timeout as failed with timeout message (#165)
Problem:

Previously, if an OpenJD Action was canceled due to a timeout being
reached then we'd report the action as just canceled. The change to
update to openjd-sessions 0.5.0
( https://github.com/casillas2/deadline-cloud-worker-agent/pull/160 )
made it so that timeout actions would report as FAILED, but didn't
change the failure message to make it clear that the reason for the
failure was a timeout.

Solution:

We mutate the action status when we recieve it to override the failure
message with one that indicates that the action has reached its runtime
limit.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>
Signed-off-by: Graeme McHale <gmchale@amazon.com>