feat!: retool scale-in behaviour
Summary:

We're modifying the worker contract as it pertains to
scale-in/service-initiated-drain operations. Currently, the behavior is
to DeleteWorker and then try to shut down the host. The new proposed
behaviour is to go to STOPPING and then repeatedly heartbeat with the
service while attempting to shutdown the host; with the expectation that
a nominally behaving host will properly shut down. The service will then
Delete the worker once the heartbeats stop.

Worker behaviour with locally initiated shutdowns (spot interruptions,
SIGTERM, SIGINT, etc) remains unchanged; we transition to STOPPING while
trying to end the work we have and then transition to STOPPED and exit
when done.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>