feat: overhaul agent logging to introduce structured logs
Problem:

The Agent's logs are in need of an overhaul. Some of the issues include:
1. Their free-form nature makes it difficult to construct log queries
   to focus in on different aspects of a run. e.g. "Find all logs
   related to Session X."
2. They are, at times, too verbose; notably, logs tend to get dominated
   by the 4x+ per minute regular queries to update the worker's
   schedule.
3. They contain information that is considered sensitive for the Agent
   log. Notably, they contain Task parameter names vand values.

Solution:

This is a pretty big change with multiple dimensions. At its core is
changing the logs to be well structured. By default, log entries are
emited as a JSON encoded structure. The Agent can be configured to emit
"human readable" logs instead, and when doing so the log entries still
follow a standard form that is intended to make them easier to
understand at a glance.

Notable additional changes include:
1. A full pass on most of the Agent's log statements to reduce
   duplication and generally try to make them more succinct.
2. API requests to deadline.UpdateWorkerSchedule are suppressed from the
   logs entirely if the request is identical to the previous request.
3. API responses from deadline.UpdateWorkerSchedule only show response
   parameters if the response differs from the previous response or is
   an Error.
4. Addition of delimiters in the Session log for job attachments' sync
   actions.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>