fix: change windows logon type from NETWORK to INTERACTIVE (#234)
Problem:

Some applications have security controls that restrict what a network
logon user can do, but do not have such restrictions for interactive
logons.

Solution:

Change the win32 logon kind from NETWORK to INTERACTIVE. NETWORK was
oridinally chosen as it was believed that it was required for the logon
to have access to network shares, but INTERACTIVE grants the same
access so this should be a safe change.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>
Signed-off-by: Brian Axelson <baxelson@amazon.com>