feat: agent logs contain more verbose API request/response information (#215)
Summary:

Our current agent log is logging the API calls that it makes for
auditing and debugging purposes. However, the API log didn't really
contain sufficient information to understand what is happening; for
example, we wouldn't know what sessionactions were assigned in an
UpdateWorkerSchedule call, or whether the service told the agent to
shutdown in a response to the same API.

Further, our logs have a content filter that assumes that all request/responses
are flat and that filtering decisions can be made only at the top-level.
That is not the case for the AWS Deadline Cloud APIs that the agent calls;
many of these have useful information throughout, but do contain some
sensitive data deeply nested within structures that the agent should
absolutely never log. This restricts the agent's ability to make fine
grained filtering decisions, and forces it to generally choose to log
very little from an API request/response.

Solution:

I reworked the API request/response filtering logic to allow for
making filtering decisions deeply within nested structures, rather than
just at the top level. I also created an ignore list where any API
request that matches the ignore list will not appear in the log. API
requests that are neither in the allow list nor ignore list are logged
with heavy redaction. Then, I updated our filters to print much more
useful information from the request/response of an API request.

Furthermore, I added the request identifier to the API response log, and
added UpdateWorkerSchedule to the filter so that customers can see that
request as well.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>