fix: return sessionactions if Session is stopped (#252)
Problem:

It's possible during error or drain flows for the Worker Agent's
"Session" main thread to have exited. Generally, the Agent will return
all pipelined sessionactions when this happens. But, it's possible that
the response to returning those sessionactions contains additional
session actions to pipeline, since the service doesn't know that the
agent hit a problem. If that happens, then the Agent currently queues up
those new sessionactions in the Session, but that Session's main thread
is dead so it'll never run them. The result is, effectively, a deadlock;
the service is waiting for the Agent to return terminal statuses for
each of the new sessionactions, but the Agent isn't going to ever run
those sessionactions.

Solution:

This is a bit of a hack, but it gets around the immediate issue. Now,
when the Agent is given new sessionactions for a Session it'll check if
the Session's main thread has exited (by checking that the future that's
running it has exited). If the future has exited, then the Agent will no
longer queue up those sessionactions, but will instead immediately
return them to the service. The first action will be marked FAILED so
that the service knows that this Session has terminated.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>