feat: adds data on action kind and queue length to logs
Request:

When looking at Agent Logs it is useful to know:

1. What kind of SessionAction is starting/ending. This helps understand
   the flow of the running Session(s) just from the agent log and can be
   a useful troubleshooting tool.
2. How many actions are queued up in a Session's internal pipeline. As
   with the above, this can be a useful troubleshooting datapoint, but
   can also be used to generate some statistics on scheduling behavior.

Solution:

This adds:

1. A 'kind' field to SessionActionLogEvents with values: TaskRun,
   EnvEnter, EnvExit, and JobAttachSyncInput.
2. A 'queue_length' field to SessionLogEvents of subtype Add and Remove.

It also adds a warning message to the start of the Agent log to inform
the customer about the expectations that they can have about log
contents and formatting. This was added because not all customers will
see the README on the GitHub repository.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>