fix: handle case where BatchGetJobEntity returns no jobRunAsUser
Problem:

The jobRunAsUser response field of the JobDetailsEntity structure within
a BatchGetJobEntity request is an optional field. This Worker Agent
currently treats it as a required field; resulting in failed Sessions
when the field is absent.

Solution:

Handle the case of missing the jobRunAsUser field in the response. It
is a service invariant that this field is always be present when the
Worker is within a CMF. When the Worker is running within an SMF, then
the field may be optional; but, further, in this case the Agent will
always be running with a local jobRunAsUser override. We fail Sessions
when this invariant is violated.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>