fix: error due to out-of-range process exit code
Problem:
On Win32 systems we may receive process exit codes that are outside of
the range of a 32-bit signed integer (as expected by UpdateWorkerAgent
API), but in range of a 32-bit unsigned integer. This causes a
validation error to be raised which the agent treats as an unrecoverable
error and exits.

Solution:
Convert all process exit codes to be in range of a 32-bit signed integer
by casting the bits to the type.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>