chore: adds cloudformation template for e2e testing resources
Summary:
We have end-to-end tests that require a very specific setup of VPC,
networking, secrets, IAM Roles, etc, but we have no information in the
GitHub repository regarding how to set that all up. This adds a
CloudFormation template that will deploy all of the required e2e testing
resources. It also adds some support scripting to deploy the template,
and get the required values from the deployed resources so that the e2e
tests can be run.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>